requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
tiktoken
transformers
torch
streamlit